+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 3.3.0 - 2023-11-15
+
+### Added
+
+- Formatting support behind the `github-code-owners.format.enabled` setting (#24). Thanks @s-weigand!
+- Tests for new formatting code (#24). Thanks @s-weigand!
+
 ## 3.2.1 - 2023-11-12
 
 ### Changed